update uid output

+++ Agora-Unreal-SDK-CPP-Example/Source/AgoraExample/AgoraAudioWidget.cpp
@@ -89,7 +89,7 @@ void UAgoraAudioWidget::onJoinChannelSuccess(const char* channel, agora::rtc::ui
 {
 	AsyncTask(ENamedThreads::GameThread, [=]()
 	{
-		GEngine->AddOnScreenDebugMessage(-1, 30.f, FColor::Blue, FString::Printf(TEXT("UAgoraAudioWidget::JoinChannelSuccess uid: %u"), uid));
+		GEngine->AddOnScreenDebugMessage(-1, 30.f, FColor::Blue, FString::Printf(TEXT("UAgoraAudioWidget::JoinChannelSuccess uid: %d"), (int64)uid));
 	});
 }
 
@@ -99,7 +99,7 @@ void UAgoraAudioWidget::onAudioVolumeIndication(const agora::rtc::AudioVolumeInf
 	{
 		for (unsigned int i = 0; i < speakerNumber; i++)
 		{
-			GEngine->AddOnScreenDebugMessage(-1, 3.f, FColor::Red, FString::Printf(TEXT("onAudioVolumeIndication uid:%lld,vad:%d,voicepatch:%d,volume %d,totalvolume:%d"), speakers[i].uid, speakers[i].vad, speakers[i].voicePitch, speakers[i].volume, totalVolume));
+			GEngine->AddOnScreenDebugMessage(-1, 3.f, FColor::Red, FString::Printf(TEXT("onAudioVolumeIndication uid:%lld,vad:%d,voicepatch:%d,volume %d,totalvolume:%d"), (int64)speakers[i].uid, speakers[i].vad, speakers[i].voicePitch, speakers[i].volume, totalVolume));
 		}
 	});
 }
@@ -108,7 +108,7 @@ void UAgoraAudioWidget::onUserJoined(agora::rtc::uid_t uid, int elapsed)
 {
 	AsyncTask(ENamedThreads::GameThread, [=]()
 	{
-		GEngine->AddOnScreenDebugMessage(-1, 30.f, FColor::Blue, FString::Printf(TEXT("UAgoraAudioWidget::onUserJoined uid: %u"), uid));
+		GEngine->AddOnScreenDebugMessage(-1, 30.f, FColor::Blue, FString::Printf(TEXT("UAgoraAudioWidget::onUserJoined uid: %d"), (int64)uid));
 	});
 
 }
